+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Add support for creating and updating PayPal customer session
 * Add support for getting PayPal customer insights
+* Bug fix for `failOnDuplicatePaymentMethodForCustomer` in `PaymentMethod`
 
 ## 3.37.0
 * Add `fail_on_duplicate_payment_method_for_customer` option to